Enable new features with config option
This change converts the Config value, which was formerly attached to
pkg, into a func passed to the template so that it can be accessed by
any sub-template. This enabled the new Consts, Vars, Factories, and
Methods config options (which all needed to be accessed by the `types`
sub-template).
