+++ testdata/pkg1/README.md
@@ -73,36 +73,6 @@ An image:
 
 ![gopher](https://golang.org/doc/gopher/frontpage.png)
 
-## Constants
-
-```golang
-const (
-    // ConstVal1 is a const in a const block.
-    ConstVal1 int = 1
-)
-```
-
-ConstVal2 is a const outside a const block.
-
-```golang
-const ConstVal2 string = "2"
-```
-
-## Variables
-
-```golang
-var (
-    // VarVal1 is a var in a var block.
-    VarVal1 int = 3
-)
-```
-
-VarVal2 is a var outside a var block.
-
-```golang
-var VarVal2 string = "4"
-```
-
 ## Sub Packages
 
 * [subpkg1](./subpkg1): Package subpkg1 is the first subpackage
