+++ testdata/pkg10_package_consts_vars/README.md
@@ -1,0 +1,33 @@
+# pkg1
+
+Package pkg1 is a testing package.
+
+## Constants
+
+```golang
+const (
+    // ConstVal1 is a const in a const block.
+    ConstVal1 int = 1
+)
+```
+
+ConstVal2 is a const outside a const block.
+
+```golang
+const ConstVal2 string = "2"
+```
+
+## Variables
+
+```golang
+var (
+    // VarVal1 is a var in a var block.
+    VarVal1 int = 3
+)
+```
+
+VarVal2 is a var outside a var block.
+
+```golang
+var VarVal2 string = "4"
+```
