+++ testdata/pkg10_package_consts_vars/goreadme.json
@@ -1,0 +1,4 @@
+{
+    "consts": true,
+    "vars": true
+}
