+++ testdata/pkg11_types_consts_vars/go.mod
@@ -1,0 +1,3 @@
+module pkg6
+
+go 1.15
