+++ testdata/pkg11_types_consts_vars/goreadme.json
@@ -1,0 +1,7 @@
+{
+    "types": true,
+    "factories": true,
+    "methods": true,
+    "consts": true,
+    "vars": true
+}
