+++ testdata/pkg8_types/pkg.go
@@ -1,26 +1,19 @@
 // Package pkg1 is a testing package.
 package pkg1
 
-const (
-	// ConstVal1 is a const in a const block.
-	ConstVal1 int = 1
-)
-
-// ConstVal2 is a const outside a const block.
-const ConstVal2 string = "2"
-
-var (
-	// VarVal1 is a var in a var block.
-	VarVal1 int = 3
-)
-
-// VarVal2 is a var outside a var block.
-var VarVal2 string = "4"
-
 // ExampleType is a type
 type ExampleType struct {
 	val              int
 	ExampleInterface interface{}
+}
+
+// ExampleTypeFactory is a factory function for ExampleType.
+func ExampleTypeFactory() ExampleType {
+	return ExampleType{1, "test"}
+}
+
+// ExampleMethod is a method on ExampleType.
+func (e ExampleType) ExampleMethod() {
 }
 
 // ExampleType2 is a type with an array
@@ -31,29 +24,3 @@ type ExampleType2 struct {
 
 // ExampleTypeInt is a one-liner type
 type ExampleTypeInt struct{ val int }
-
-// ExampleFactoryFunction is a function that returns an ExampleType by value.
-func ExampleFactoryFunction() ExampleType {
-	return ExampleType{
-		val:              1,
-		ExampleInterface: "something",
-	}
-}
-
-// ExampleFactoryFunction2 is a function that returns an ExampleType by pointer, and an error.
-func ExampleFactoryFunction2() (*ExampleType, error) {
-	return &ExampleType{
-		val:              1,
-		ExampleInterface: "something",
-	}, nil
-}
-
-// ExampleMethod is a method on an ExampleType that takes the receiver by value.
-func (et ExampleType) ExampleMethod() string {
-	return "something"
-}
-
-// ExampleMethod2 is a method on an ExampleType that takes the receiver by pointer.
-func (et *ExampleType) ExampleMethod2() string {
-	return "*something"
-}
